Refactor database models and services to use uint for IDs
- Updated ID types from int to uint in models: Book, Category, Chapter, Favorite, Source, User.
- Adjusted corresponding service methods to accept uint for ID parameters.
- Modified database queries to align with the new ID type.
- Created initial SQL migration script for setting up the database schema with the updated models.

+++ server/api/book_handlers.go
@@ -25,7 +25,7 @@ func (api *api) ListBooks(r web.Request) (any, error) {
 		sourceID = 0 // default sourceID
 	}
 
-	books, total, err := api.s.GetBooks(page, limit, sourceID)
+	books, total, err := api.s.GetBooks(page, limit, uint(sourceID))
 	if err != nil {
 		return models.BooksPaginatedResponse{}, err
 	}
@@ -47,5 +47,5 @@ func (api *api) GetBook(r web.Request) (any, error) {
 		return nil, err
 	}
 
-	return api.s.GetBookByID(id)
+	return api.s.GetBookByID(uint(id))
 }
